requirements.txt
# Databricks SDK for workspace and API interactions
databricks-sdk>=0.18.0

# Environment variable management
python-dotenv>=1.0.0

# Microsoft Authentication Library for Azure AD authentication (recommended for Databricks)
msal>=1.24.0

# HTTP requests for OAuth token handling
requests>=2.28.0

# Streamlit for web app interface
streamlit>=1.28.0
